\name{cy}
\alias{cy}
\docType{data}
\title{
Example of a functional data set
}
\description{
Example of a functional data set	
}
\usage{data(cy)}
\format{
  A data frame with 250 observations on the following 3 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Time}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
  }
}

\keyword{datasets}
