% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_pcf}
\alias{extract_pcf}
\title{extract principal component function from a fitted object}
\usage{
extract_pcf(nharm, fit.obj)
}
\arguments{
\item{nharm}{integer corresponding to the principal component function. nharm = 1 corresponds to the leadng principal component function.}

\item{fit.obj}{fitted object from output of estimate_eigenfunctions() from which PCFs will be derived.}
}
\value{
the principal component function corresponding to nharm
}
\description{
extract principal principal component funciton from a fitted object.
}

